Update README.md
Keep snapshots forever unless instructed otherwise to avoid accidental deletion in case the d flag is missing. Better to be explicit about what to delete. This is specially important for compliance reasons. For example In the USA, the IRS financial record requirements are 7 years. Deleting a snapshot by mistake might result in a real disaster.